Updated to use hzeller/rpi-rgb-led-matrix to drive the matrix.
