+++ .gitignore
@@ -1,3 +1,5 @@
 desktop.ini
 Thumbs.db
 winscp.rnd
+/__pycache__
+/pi.ppm
