+++ RunClock.sh
@@ -3,4 +3,4 @@
 ps aux | grep -ie PiClock | awk '{print "sudo kill -9 " $2}' | sh -x
 #ps aux | grep -ie RunClock | awk '{print "sudo kill -9 " $2}' | sh -x
 
-sudo python ~/PiClock/PiClock.py
+sudo python3 ~/PiClock/PiClock.py
